+++ surf-flutter-courses-template-main/lib/color_palette/support/color_palette_cell_widget.dart
